Verify the history and triage schema against Postgres, and correct a comment
The migration and the eight new endpoints had never been run against a
database. They now have been, against PostgreSQL 18.4:

- 000025 applies on top of all 24 prior migrations, and again via the
  server's own embedded golang-migrate path on a fresh database.
- The down migration drops cleanly and the up re-applies afterwards.
- Every endpoint answers correctly over seeded history: flip counts,
  failing_since_commit, amnesty's rate check, and the accuracy precision
  and false-green counts.
- Re-posting a verdict upserts rather than appends, including for
  suite-level rows where external_test_id is NULL and for rows where both
  key columns are NULL. NULLS NOT DISTINCT and ON CONFLICT arbiter
  inference do work together, which was the part most likely not to.
- The correction-pairing, confidence-range, and verdict-enum checks all
  reject what they should.

That run also disproved the reason given for the casts added here. pgx
v5.9.1 scans an uncast count() and an uncast sum(bigint) into Go int and
int64 without error, so the history endpoint was never going to fail on its
first row, and the flakiness aggregate below — which leaves its own count()s
uncast — returns correct numbers as written.

The casts are kept because they make the column types read as what the Go
struct declares, but the comment claiming they are load-bearing is replaced:
left as it was, it invited someone to "fix" the uncast aggregate that is
already fine, and count(*)::int in fact narrows a bigint that Go would
otherwise hold safely.

+++ apps/server/internal/api/testhistory/handlers.go
@@ -42,8 +42,9 @@ const (
 // groupRollupSQL is the shared CTE: every report group that ran the requested
 // test, with the test's per-shard rows collapsed to a single outcome.
 //
-// Args, in order: $1 external_test_id, $2 repository, $3 branch (”=any),
-// $4 framework (”=any), $5 run_group (”=any), $6 since (zero time = no bound).
+// Args, in order: $1 external_test_id, $2 repository, $3 branch, $4 framework,
+// $5 run_group, $6 since. The three filter args treat an empty string as "any";
+// since treats NULL as unbounded.
 const groupRollupSQL = `
 	WITH matched AS (
 		SELECT g.id, g.commit_sha, g.gh_run_id, g.gh_pr_number, g.branch,
@@ -61,8 +62,14 @@ const groupRollupSQL = `
 	),
 	rolled AS (
 		SELECT id, commit_sha, gh_run_id, gh_pr_number, branch, name, run_group, created_at,
-		       count(*)                                                        AS shard_rows,
-		       sum(coalesce(duration_ms, 0))                                   AS duration_ms,
+		       -- count() returns bigint and sum(bigint) returns numeric. Both
+		       -- casts are cosmetic, not load-bearing: pgx v5 scans either into
+		       -- the Go int/int64 these land in without them (verified against
+		       -- PG 18.4), which is why the flakiness aggregate below leaves its
+		       -- own count()s uncast. Kept only so the column types read as what
+		       -- the Go struct declares.
+		       count(*)::int                                                   AS shard_rows,
+		       sum(coalesce(duration_ms, 0))::bigint                           AS duration_ms,
 		       bool_or(status IN ('passed', 'flaky'))                          AS ever_passed,
 		       bool_or(status IN ('failed', 'timedOut', 'interrupted'))        AS ever_failed
 		FROM matched
